% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_to_dt.R
\name{med_to_dt}
\alias{med_to_dt}
\title{Convert MedPC File to an R data table}
\usage{
med_to_dt(file, group = NA, condition = NA)
}
\arguments{
\item{file}{string or vector; complete path to MedPC data file}

\item{group}{string (optional); experimental group of animal}

\item{condition}{string (optional); experimental condition (i.e. drug)}
}
\value{
data table with the following variables: \cr
 subject \cr
 group \cr
 condition \cr
 date = ddmmyy \cr
 trial = trial number \cr
 pLeftBlink = p(left blink) on trial \cr
 pRightBlink = p(right blink) on trial, 1-p(left blink) \cr
 blinkSequence = actual sequence of blinks from start of trial until response \cr
 choice = choice on trial; 0=left, 1=right
 RT = response time for trial
 omission = did rat get reward (only applicable for correct trials)? 0 if yes, 1 if no
 rewardRT = response time from decision to reward; 0 if incorrect
 startTime = time in session at start of trial
 decisionTime = time in session at decision
 rewardTime = time in session at reward receipt
 trialTime = duration of trial
 pCorrectBlink = p(blink) on correct side; p(left blink) if left is correct, p(right blink) if right is correct
 correctSide = 0 is left is correct answer, 1 if right is correct answer
 correctChoice = 0 if incorrect, 1 if correct
}
\description{
Convert MedPC File to an R data table
}
