% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_blinks_data.R
\name{save_blinks_data}
\alias{save_blinks_data}
\title{Save data frame of foraging data to a file}
\usage{
save_blinks_data(dat, path = getwd(), name = "", as_csv = F)
}
\arguments{
\item{dat}{data frame or list of data frames created by med_to_dt}

\item{path}{string (optional); complete path to save directory, if not given, uses current working directory}

\item{name}{string (optional); base name of file. if not provided, default is 'subject_date' of first row}

\item{as_csv}{logical; if true, save as .csv file. if false, save as .Rdata file with a data frame called 'forage_data'. default=false}
}
\value{
complete path to saved file
}
\description{
Save data frame of foraging data to a file
}
